refactor(zip-compressor): ts-node化
