+++ README.md
@@ -1,0 +1,9 @@
+## get.wwaing.com
+
+WWA Wing 配布物ページ と 配布物生成スクリプト
+
+## TODO 
+- [ ] WWAWing リポジトリ の 自動 publish
+- [ ] エンジンバージョン取得 from `@wwawing/engine` の　`package.json`
+- [ ] `zip_compressor/output/*.zip` を `docs` にバージョンつきのファイル名にしてコピーするスクリプトを書く
+- [ ] `npm start` で全部走るようにする
