+++ package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@wwawing/get-wwawing-com",
+  "version": "0.0.0",
+  "description": "WWA Wing Release Pages",
+  "main": "index.js",
+  "directories": {
+    "doc": "docs"
+  },
+  "scripts": {
+    "make-dist": "npm run clean:dist && ts-node ./scripts/zip-compressor/make-wwawing-dist && ts-node ./scripts/zip-compressor/make-zip",
+		"clean": "npm run clean:dist",
+    "clean:dist": "shx rm -rf ./scripts/zip-compressor/output",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/WWAWing/get.wwawing.com.git"
+  },
+  "author": "WWA Wing Team",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/WWAWing/get.wwawing.com/issues"
+  },
+  "homepage": "https://github.com/WWAWing/get.wwawing.com#readme",
+  "devDependencies": {
+    "@types/jszip": "^3.1.6",
+    "@types/node": "^12.0.8",
+    "@types/shelljs": "^0.8.5",
+    "glob": "^7.1.4",
+    "jszip": "^3.2.1",
+    "lerna": "^3.15.0",
+    "pug": "^2.0.3",
+    "shelljs": "^0.8.3",
+    "shx": "^0.3.2",
+    "ts-node": "^8.3.0",
+    "typescript": "^3.5.2"
+  }
+}
