--- packages/generate-release-page/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "@wwawing/generate-release-page",
-  "version": "1.0.0",
-  "description": "",
-  "main": "generate-page.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "WWA Wing Team",
-  "license": "MIT",
-  "devDependencies": {
-    "pug": "^2.0.3"
-  }
-}
