--- packages/zip-compressor/.gitignore
@@ -1,3 +1,0 @@
-lib/
-output/
-node_modules
