--- packages/zip-compressor/package.json
@@ -1,37 +1,0 @@
-{
-	"name": "@wwawing/build-utils",
-	"version": "0.0.0",
-	"description": "build utilies for WWA Wing",
-	"main": "lib/make-wwapage.js",
-	"directories": {
-		"lib": "lib"
-	},
-	"scripts": {
-		"build": "npm run clean:script && tsc -p .",
-		"make-dist": "npm run clean:dist && node ./lib/make-wwawing-dist && node ./lib/make-zip",
-		"clean": "npm run clean:script && npm run clean:dist",
-		"clean:script": "shx rm -rf lib",
-		"clean:dist": "shx rm -rf output",
-		"test": "echo \"Error: no test specified\" && exit 1"
-	},
-	"repository": {
-		"type": "git",
-		"url": "git+https://github.com/WWAWing/WWAWing.git"
-	},
-	"author": "WWA Wing Team",
-	"license": "MIT",
-	"devDependencies": {
-		"@types/jszip": "^3.1.4",
-		"@types/node": "^10.5.1",
-		"@types/shelljs": "^0.8.0",
-		"@wwawing/engine": "3.2.0",
-		"@wwawing/assets": "0.0.0",
-		"@wwawing/loader": "0.0.0",
-		"@wwawing/debug-server": "0.0.0",
-		"glob": "^7.1.2",
-		"jszip": "^3.1.5",
-		"shelljs": "^0.8.2",
-		"shx": "^0.3.2",
-		"typescript": "^2.9.2"
-	}
-}
