--- packages/zip-compressor/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "compilerOptions": {
-    "rootDir": "./src",
-    "outDir": "./lib",
-    "target": "es5",
-    "module": "commonjs",
-    "moduleResolution": "node",
-    "lib": [
-      "es2017",
-      "es2016",
-      "es2015",
-      "es5",
-      "dom"
-    ],
-    "strict": true,
-    "noLib": false,
-    "declaration": true,
-    "removeComments": true,
-    "experimentalDecorators": true,
-    "noUnusedParameters": false,
-    "noUnusedLocals": true
-  },
-  "include": [
-    "src/**/*.ts"
-  ]
-}
