+++ scripts/generate-release-page/index.pug
@@ -1,0 +1,32 @@
+doctype html
+html(lang="ja")
+    head
+        meta(charset="UTF-8")
+        title WWA Wing 配布ページ
+    body
+        h1 WWA Wing 配布ページ
+        h2 最新版
+        ul
+            li
+                a(href="./wwawing-dist-latest.zip") 完全版 
+            li
+                a(href="./wwawing-update-latest.zip") 更新版 
+
+        h2 過去のリリース
+        ul
+            each version in releases
+                li
+                    | #{version}
+                    | (
+                    a(href="./wwawing-dist-" + version + ".zip") 完全版
+                    | )
+                    | (
+                    a(href="./wwawing-update-" + version + ".zip") 更新版
+                    | )
+        footer
+            p
+                a(href="./old-releases") さらに過去のリリースを見る
+            p
+                a(href="https://wwawing.com/") WWA Wingのトップに戻る
+            p
+                | (C) 2013-2018 WWA Wing Team
